refactor(admin-view): split template emitters
