+++ .sampo/changesets/913-split-admin-view-templates.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split admin-view template emitters into focused default, REST, core-data, settings, and shared modules while keeping the existing public template import path compatible.
